HelpCenter: minor restructuring for querying support tickets
