+++ packages/data-stores/src/support-queries/types.ts
@@ -27,3 +27,13 @@ export interface OtherSupportAvailability {
 	is_user_eligible_for_directly: boolean;
 	is_user_eligible_for_chat: boolean;
 }
+
+export interface SupportTicket {
+	id?: number;
+	status: string;
+	subject: string;
+	time: Date;
+	type: string;
+	url: string;
+	when: string;
+}
