+++ packages/data-stores/src/support-queries/use-active-support-tickets-query.ts
@@ -1,15 +1,10 @@
-/* eslint-disable no-restricted-imports */
 import { useQuery } from 'react-query';
-import { useSelector } from 'react-redux';
 import wpcomRequest from 'wpcom-proxy-request';
-import { getCurrentUserEmail } from 'calypso/state/current-user/selectors';
-import { SupportTicket } from '../types';
+import { SupportTicket } from './types';
 
 const ACTIVE_STATUSES = [ 'New', 'Open', 'Hold' ];
 
-export const useActiveSupportTicketsQuery = ( queryOptions = {} ) => {
-	const email = useSelector( getCurrentUserEmail );
-
+export const useActiveSupportTicketsQuery = ( email: string, queryOptions = {} ) => {
 	return useQuery< SupportTicket[] >(
 		[ 'activeSupportTickets', email ],
 		async (): Promise< SupportTicket[] > => {
