+++ packages/help-center/src/components/help-center-contact-page.tsx
@@ -4,23 +4,22 @@
  */
 import { recordTracksEvent } from '@automattic/calypso-analytics';
 import { Spinner } from '@automattic/components';
-import { Icon, comment } from '@wordpress/icons';
+import { useActiveSupportTicketsQuery } from '@automattic/data-stores';
+import { comment, Icon } from '@wordpress/icons';
 import { useI18n } from '@wordpress/react-i18n';
 import classnames from 'classnames';
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { Link, LinkProps } from 'react-router-dom';
+import { getCurrentUserEmail } from 'calypso/state/current-user/selectors';
 import { getSectionName } from 'calypso/state/ui/selectors';
 /**
  * Internal Dependencies
  */
 import { BackButton } from '..';
-import {
-	useActiveSupportTicketsQuery,
-	useShouldRenderChatOption,
-	useShouldRenderEmailOption,
-	useStillNeedHelpURL,
-} from '../hooks';
+import { useShouldRenderChatOption } from '../hooks/use-should-render-chat-option';
+import { useShouldRenderEmailOption } from '../hooks/use-should-render-email-option';
+import { useStillNeedHelpURL } from '../hooks/use-still-need-help-url';
 import Mail from '../icons/mail';
 import { HelpCenterActiveTicketNotice } from './help-center-notice';
 import { SibylArticles } from './help-center-sibyl-articles';
@@ -37,7 +36,8 @@ export const HelpCenterContactPage: React.FC = () => {
 
 	const renderEmail = useShouldRenderEmailOption();
 	const renderChat = useShouldRenderChatOption();
-	const { data: tickets, isLoading: isLoadingTickets } = useActiveSupportTicketsQuery();
+	const email = useSelector( getCurrentUserEmail );
+	const { data: tickets, isLoading: isLoadingTickets } = useActiveSupportTicketsQuery( email );
 
 	if ( renderChat.isLoading || isLoadingTickets ) {
 		return (
