--- packages/help-center/src/hooks/index.ts
@@ -1,4 +1,0 @@
-export { useActiveSupportTicketsQuery } from './use-active-support-tickets-query';
-export { useStillNeedHelpURL } from './use-still-need-help-url';
-export { useShouldRenderEmailOption } from './use-should-render-email-option';
-export { useShouldRenderChatOption } from './use-should-render-chat-option';
